Migrate to ejs partials and de-governmentize
EJS partials allow us to separate out concerns and that way forks
of this project only have to modify the partials that impact them.

Remove as many instances to "government" as possible so that it is
more of an example which can be forked for other projects.

Remove absolute paths to analytics.usa.gov and instead use
paths to this server.

Because of the use of GMT which may be ahead of or behind the
current time, use the `last-48-hours` report instead of the
`today` report so that at least 24 hours of data can be rendered.

For 7/30 day views, use top pages instead of top domains.  This makes
it more consistent with the realtime view, so that all 3 views
are essentially showing the same data over 3 time periods.

Comment out the government's DAP analytics program.  Add it back in
with your fork if you need it.